requirements.txt
streamlit
pillow
torch
torchvision
pytorch-grad-cam 